pyproject.toml
[tool.poetry]
name = "ig_bot"
version = "0.1.0"
description = "Instagram video/photo downloader bot for Telegram"
authors = ["Danil Ineev <sai@notreally.media>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.1"
instaloader = "^4.5.5"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
